requirements.txt
requests==2.31.0
pandas==2.2.1
python-dotenv==1.0.1
boto3==1.34.39
botocore==1.34.39
fastapi==0.95.0
uvicorn==0.22.0
pytest==7.4.0
httpx==0.24.1
python-multipart==0.0.9
aiofiles==23.2.1